chore(storage): restructure modules
