+++ src/storage/secondary/block/primitive_nullable_block_builder.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use super::encode::PrimitiveFixedWidthEncode;
+use super::super::PrimitiveFixedWidthEncode;
 use super::BlockBuilder;
 
 /// Encodes fixed-width data into a block, with null element support.
